Fix collections DeprecationWarning.
Since Python 3.3, the Abstract Base Classes for the collections
package have been available on the collections.abc module,
and since Python 3.7 loading them from the main module path
has been deprecated.

Update the import to use the new path, silencing the warning
and avoiding errors with future Python releases.

This doesn't work with Python 2, but that doesn't seem to be
supported. If necessary, support can be added by wrapping the
import in a try block:

    try:
        from collections.abc import Callable
    except ImportError:
        # Python 2
        from collections import Callable

That will fall back to the old method.