Fix invalid escape sequence warning.
Python 3.7 warns about unrecognized escape sequences in strings.

Remove the backslash escapes for the `/` character. This doesn't
need to be escaped for re match patterns; it only has special
meaning in sed-style substitution patterns.

Another possible fix is to use raw strings, r"^(https?:)?\/\/"
which leave the backslash characters to the re package to
interpret. It doesn't mind the sequences, but they're still
unnecessary.